feat: Add beta registration modal
Introduces a new modal for users to register for beta testing. This modal is triggered when a user clicks the "Buy" button for a gift. It includes a form for collecting user details and a success state. The `GiftDetailsModal` has been updated to manage the state and rendering of this new modal.